rename snarky -> bindings
