+++ flake.nix
@@ -251,7 +251,7 @@
                 ./dune-project
                 ./.prettierrc.cjs
                 ./src/build
-                ./src/snarky.d.ts
+                ./src/bindings.d.ts
               ];
             });
           inherit (inputs.mina.devShells."${system}".default)
