+++ src/lib/mina/v1/local-blockchain.ts
@@ -1,7 +1,7 @@
 import { SimpleLedger } from './transaction-logic/ledger.js';
 import { Ml } from '../../ml/conversion.js';
 import { transactionCommitments } from '../../../mina-signer/src/sign-zkapp-command.js';
-import { Ledger, Test, initializeBindings } from '../../../snarky.js';
+import { Ledger, Test, initializeBindings } from '../../../bindings.js';
 import { Field } from '../../provable/wrapped.js';
 import { UInt32, UInt64 } from '../../provable/int.js';
 import { PrivateKey, PublicKey } from '../../provable/crypto/signature.js';
