+++ src/lib/ml/consistency.unit-test.ts
@@ -1,4 +1,4 @@
-import { Test } from '../../snarky.js';
+import { Test } from '../../bindings.js';
 import { Random, test } from '../testing/property.js';
 import { Field, Bool } from '../provable/wrapped.js';
 import { PrivateKey, PublicKey } from '../provable/crypto/signature.js';
