+++ src/mina-signer/src/transaction-hash.unit-test.ts
@@ -1,4 +1,4 @@
-import { Test } from '../../snarky.js';
+import { Test } from '../../bindings.js';
 import {
   Common,
   hashPayment,
