+++ tsconfig.node.json
@@ -2,7 +2,7 @@
   "extends": "./tsconfig.json",
   "include": [
     "./src/index.ts",
-    "./src/snarky.js",
+    "./src/bindings.js",
     "./src/mina-signer/src",
     "./src/mina-signer/mina-signer.ts",
     "./src/js_crypto",
