+++ tsconfig.test.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./src/**/*.unit-test.ts", "./src/lib/**/*.ts", "./src/snarky.js"],
+  "include": ["./src/**/*.unit-test.ts", "./src/lib/**/*.ts", "./src/bindings.js"],
   "compilerOptions": {
     "outDir": "dist/node"
   }
