+++ tsconfig.web.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./src/index.ts", "./src/snarky.js", "./src/**/*.web.ts"],
+  "include": ["./src/index.ts", "./src/bindings.js", "./src/**/*.web.ts"],
   "compilerOptions": {
     "outDir": "dist/web"
   }
